Moved generic metrics to another module
