+++ apf/metrics/__init__.py
@@ -1,2 +1,5 @@
+from . import generic
+from .generic import *
+
 from . import kafka
 from .kafka import *
